+++ esm/index.js
@@ -36,8 +36,10 @@ export { LightTooltip, DarkTooltip } from './components/Tooltip';
 export { BottomFixedArea } from './components/BottomFixedArea';
 export { MessageScreen } from './components/MessageScreen';
 export { Calendar } from './components/Calendar';
+export { DatePicker } from './components/DatePicker';
 export { IndexNav } from './components/IndexNav';
 export { FormGroup } from './components/FormGroup';
+export { SegmentedControl } from './components/SegmentedControl';
 // themes
 export { createTheme } from './themes/createTheme';
 export { ThemeProvider } from './themes/ThemeProvider';
@@ -46,5 +48,7 @@ export { defaultInteraction } from './themes/createInteraction';
 export { defaultFrame } from './themes/createFrame';
 export { defaultSize } from './themes/createSize';
 // constants
-export { FONT_FAMILY } from './constants';
+export { FONT_FAMILY, VISUALLY_HIDDEN_STYLE } from './constants';
+// utils
+export { SequencePrefixIdProvider } from './hooks/useId';
 //# sourceMappingURL=index.js.map