+++ esm/types/props.js
@@ -1,1 +1,2 @@
+export {};
 //# sourceMappingURL=props.js.map