+++ lib/hooks/useId.d.ts
@@ -1,0 +1,5 @@
+import { FC, ReactNode } from 'react';
+export declare function useId(defaultId?: string): string;
+export declare const SequencePrefixIdProvider: FC<{
+    children: ReactNode;
+}>;