+++ src/components/Icon/Icon.tsx
@@ -310,7 +310,7 @@ export interface IconProps {
   size?: number
 }
 
-export interface Props extends IconProps {
+export interface Props extends IconProps, React.SVGAttributes<SVGAElement> {
   name: keyof typeof iconMap
   className?: string
 }
@@ -621,7 +621,7 @@ export const iconMap = {
   'fa-yen-sign': FaYenSign,
 }
 
-export const Icon: React.FC<Props> = ({ name, className = '', ...props }) => {
+export const Icon: React.FC<Props> = ({ name, className = '', role = 'img', ...props }) => {
   const SvgIcon = iconMap[name]
-  return <SvgIcon className={className} {...props} />
+  return <SvgIcon className={className} role={role} {...props} />
 }