+++ src/index.ts
@@ -75,8 +75,10 @@ export { LightTooltip, DarkTooltip } from './components/Tooltip'
 export { BottomFixedArea } from './components/BottomFixedArea'
 export { MessageScreen } from './components/MessageScreen'
 export { Calendar } from './components/Calendar'
+export { DatePicker } from './components/DatePicker'
 export { IndexNav, IndexNavItemProps } from './components/IndexNav'
 export { FormGroup } from './components/FormGroup'
+export { SegmentedControl, SegmentedControlOption } from './components/SegmentedControl'
 
 // themes
 export { createTheme } from './themes/createTheme'
@@ -87,4 +89,7 @@ export { defaultFrame } from './themes/createFrame'
 export { defaultSize } from './themes/createSize'
 
 // constants
-export { FONT_FAMILY } from './constants'
+export { FONT_FAMILY, VISUALLY_HIDDEN_STYLE } from './constants'
+
+// utils
+export { SequencePrefixIdProvider } from './hooks/useId'